chore(sn): move Peer to root of sn
This will let us use the same struct across routing and clients
