+++ sn/src/lib.rs
@@ -43,6 +43,7 @@ extern crate tracing;
 
 pub mod client;
 mod dbs;
+mod peer;
 
 #[cfg(test)]
 /// Helpers for analysis of testnet logs
