+++ sn/src/routing/core/api.rs
@@ -10,6 +10,7 @@ use super::{delivery_group, Comm, Core};
 use crate::dbs::UsedSpace;
 use crate::messaging::system::{JoinResponse, MembershipState, NodeState, SigShare, SystemMsg};
 use crate::messaging::WireMsg;
+use crate::peer::Peer;
 use crate::routing::{
     core::Proposal,
     error::Result,
@@ -17,7 +18,7 @@ use crate::routing::{
     network_knowledge::{NetworkKnowledge, SectionAuthorityProvider, SectionKeyShare},
     node::Node,
     routing_api::command::Command,
-    Event, Peer,
+    Event,
 };
 use secured_linked_list::SecuredLinkedList;
 use std::{collections::BTreeSet, net::SocketAddr, path::PathBuf};
