+++ sn/src/routing/core/delivery_group.rs
@@ -8,10 +8,11 @@
 
 use crate::elder_count;
 use crate::messaging::DstLocation;
+use crate::peer::Peer;
 use crate::routing::{
     error::{Error, Result},
     network_knowledge::NetworkKnowledge,
-    supermajority, Peer,
+    supermajority,
 };
 use itertools::Itertools;
 use std::{cmp, iter};
