+++ sn/src/routing/core/msg_handling/mod.rs
@@ -17,13 +17,14 @@ mod service_msgs;
 mod update_section;
 
 use super::Core;
+use crate::peer::{Peer, UnnamedPeer};
 use crate::routing::{
     log_markers::LogMarker,
     messages::{NodeMsgAuthorityUtils, WireMsgUtils},
     network_knowledge::{NetworkKnowledge, SectionPeers},
     relocation::RelocateState,
     routing_api::command::Command,
-    Error, Event, MessageReceived, Peer, Result, UnnamedPeer, MIN_LEVEL_WHEN_FULL,
+    Error, Event, MessageReceived, Result, MIN_LEVEL_WHEN_FULL,
 };
 use crate::types::{Chunk, Keypair, PublicKey};
 use crate::{
