+++ sn/src/routing/core/msg_handling/proposals.rs
@@ -9,7 +9,8 @@
 use super::Core;
 
 use crate::messaging::{signature_aggregator::Error as AggregatorError, MessageId};
-use crate::routing::{core::Proposal, dkg::SigShare, routing_api::command::Command, Peer, Result};
+use crate::peer::Peer;
+use crate::routing::{core::Proposal, dkg::SigShare, routing_api::command::Command, Result};
 
 // Decisions
 impl Core {
