+++ sn/src/routing/core/msg_handling/relocation.rs
@@ -12,13 +12,14 @@ use crate::messaging::{
     system::{RelocateDetails, RelocatePromise, SystemMsg},
     AuthorityProof, SectionAuth,
 };
+use crate::peer::Peer;
 use crate::routing::{
     core::{bootstrap::JoiningAsRelocated, Proposal},
     error::Result,
     log_markers::LogMarker,
     relocation::{self, RelocateAction, RelocateDetailsUtils, RelocateState},
     routing_api::command::Command,
-    Event, Peer,
+    Event,
 };
 use xor_name::XorName;
