+++ sn/src/routing/core/msg_handling/resource_proof.rs
@@ -8,12 +8,13 @@
 
 use super::Core;
 use crate::messaging::system::{JoinResponse, ResourceProofResponse, SystemMsg};
+use crate::peer::Peer;
 use crate::routing::{
     core::{RESOURCE_PROOF_DATA_SIZE, RESOURCE_PROOF_DIFFICULTY},
     ed25519,
     log_markers::LogMarker,
     routing_api::command::Command,
-    Error, Peer, Result,
+    Error, Result,
 };
 use ed25519_dalek::Verifier;
 use xor_name::XorName;
