+++ sn/src/routing/mod.rs
@@ -25,16 +25,15 @@
 // ############################################################################
 pub(crate) use self::{
     core::MIN_LEVEL_WHEN_FULL,
-    network_knowledge::{
-        peer::UnnamedPeer, section_keys::SectionKeyShare, SectionAuthorityProvider,
-    },
+    network_knowledge::{section_keys::SectionKeyShare, SectionAuthorityProvider},
 };
+use crate::peer::Peer;
+
 pub use self::{
     dkg::SectionAuthUtils,
     error::{Error, Result},
-    network_knowledge::{
-        node_state::{FIRST_SECTION_MAX_AGE, FIRST_SECTION_MIN_AGE, MIN_ADULT_AGE, MIN_AGE},
-        peer::Peer,
+    network_knowledge::node_state::{
+        FIRST_SECTION_MAX_AGE, FIRST_SECTION_MIN_AGE, MIN_ADULT_AGE, MIN_AGE,
     },
     routing_api::{
         config::Config,
