+++ sn/src/routing/network_knowledge/elder_candidates.rs
@@ -6,7 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use super::peer::Peer;
+use crate::peer::Peer;
 use std::collections::BTreeMap;
 use xor_name::{Prefix, XorName};
