+++ sn/src/routing/network_knowledge/section_peers.rs
@@ -7,10 +7,8 @@
 // permissions and limitations relating to use of the SAFE Network Software.
 
 use crate::messaging::system::{MembershipState, SectionAuth, SectionPeers as SectionPeersMsg};
-use crate::routing::{
-    network_knowledge::{NodeState, SectionAuthorityProvider},
-    Peer,
-};
+use crate::peer::Peer;
+use crate::routing::network_knowledge::{NodeState, SectionAuthorityProvider};
 use dashmap::{mapref::entry::Entry, DashMap};
 use itertools::Itertools;
 use std::{
