+++ sn/src/routing/node.rs
@@ -6,7 +6,7 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::routing::Peer;
+use crate::peer::Peer;
 use crate::types::PublicKey;
 use ed25519_dalek::Keypair;
 use std::{
